Move Pod*Exceptions to separate module
